Revert "test: bring cli test back"
