+++ lib/common/utils.js
@@ -13,29 +13,6 @@ export const waitFor = function waitFor(ms) {
   return new Promise(resolve => setTimeout(resolve, ms || 0))
 }
 
-/**
- * Prepare an object to pass to the createSportsSelectionView function
- * @param {Function} condition return true to stop the waiting
- * @param {Number} duration seconds totally wait
- * @param {Number} interval milliseconds interval to check the condition
- *
- * @returns {Boolean} true: timeout, false: condition becomes true within total time
- */
-export const waitUntil = async function waitUntil(condition, duration = 20, interval = 250) {
-  let iterator = 0
-  const steps = Math.floor(duration * 1000 / interval)
-
-  while (!condition() && iterator < steps) {
-    await waitFor(interval)
-    iterator++
-  }
-
-  if (iterator === steps) {
-    return true
-  }
-  return false
-}
-
 async function promiseFinally(fn, finalFn) {
   let result
   try {
