--- test/fixtures/cli/pages/index.vue
@@ -1,3 +1,0 @@
-<template>
-  <div>CLI Test</div>
-</template>
